Updating to latest FlutterFlow output.

+++ lib/components/template_tile_component/template_tile_component_widget.dart
@@ -167,7 +167,32 @@ class _TemplateTileComponentWidgetState
                 hoverColor: Colors.transparent,
                 highlightColor: Colors.transparent,
                 onTap: () async {
-                  await widget.component!.reference.delete();
+                  var confirmDialogResponse = await showDialog<bool>(
+                        context: context,
+                        builder: (alertDialogContext) {
+                          return AlertDialog(
+                            title: const Text('Delete component'),
+                            content: const Text(
+                                'Are you sure you want to delete the component'),
+                            actions: [
+                              TextButton(
+                                onPressed: () =>
+                                    Navigator.pop(alertDialogContext, false),
+                                child: const Text('Cancel'),
+                              ),
+                              TextButton(
+                                onPressed: () =>
+                                    Navigator.pop(alertDialogContext, true),
+                                child: const Text('Confirm'),
+                              ),
+                            ],
+                          );
+                        },
+                      ) ??
+                      false;
+                  if (confirmDialogResponse) {
+                    await widget.component!.reference.delete();
+                  }
                 },
                 child: Icon(
                   Icons.delete_forever,
